@@ -1,14 +1,11 @@
 {
-  "version": "0.1.2",
+  "version": "0.1.3",
   "license": "MIT",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "files": [
     "dist"
   ],
-  "engines": {
-    "node": ">=14"
-  },
   "scripts": {
     "start": "tsdx watch",
     "build": "tsdx build",
@@ -61,19 +58,18 @@
     }
   ],
   "devDependencies": {
-    "@size-limit/preset-small-lib": "^4.10.1",
-    "@types/node": "^14.14.35",
-    "@types/uuid": "^8.3.0",
-    "husky": "^5.2.0",
-    "size-limit": "^4.10.1",
+    "@size-limit/preset-small-lib": "^6.0.4",
+    "@types/uuid": "^8.3.1",
+    "husky": "^7.0.4",
+    "size-limit": "^6.0.4",
     "tsdx": "^0.14.1",
-    "tslib": "^2.1.0",
-    "typescript": "^4.2.3"
+    "tslib": "^2.3.1",
+    "typescript": "^4.4.4"
   },
   "dependencies": {
-    "@apollo/client": "^3.4.0-beta.19",
-    "rxjs": "^7.0.0-beta.13",
-    "uuid": "^3.4.0",
+    "@apollo/client": "^3.4.17",
+    "rxjs": "^7.4.0",
+    "uuid": "^8.3.2",
     "zen-observable": "^0.8.15"
   }
 }
