@@ -1,6 +1,6 @@
-export { AbsintheLink } from "./absinthe/apollo-link";
-export { AbsintheSubscription } from "./absinthe/subscription";
+export { AbsintheLink } from './absinthe/apollo-link';
+export { AbsintheSubscription } from './absinthe/subscription';
 
-export { PhoenixChannel } from './phoenix/channel'
-export { PhoenixSerializer } from './phoenix/serializer'
-export { PhoenixSocket } from './phoenix/socket'
+export { PhoenixChannel } from './phoenix/channel';
+export { PhoenixSerializer } from './phoenix/serializer';
+export { PhoenixSocket } from './socket/socket';
